Move intermediary version elimination to its own file
Extract eliminatingIntermediaryVersions and its helpers (findHead,
intermediariesAboveMaxVersion, rebuildWithoutEliminated, skipLookGood,
versionFromEntryName, greater) out of filtered_collector.go into
eliminating_intermediary_versions.go. Pure relocation; no behavior change.
